+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [Unreleased] - 2025-11-11
+
+### Changed
+- Updated Lucide icons from unknown to 0.553.0
+- Added 6 new icon(s)
+
+
 
 ## [v0.2.0] - 2025-11-03
 ### Changed